Added Github Actions, corrected unit tests
These changes primarily added Github actions and fixed the unit tests.
They also included the removal of benchmarks which don't make sense for
this library and I also fixed the badges in the README to properly show
the new github action builds as well as point to the new go docs site.

funnel_test_mocks.go was changed to funnel_mocks_test.go so that the
linter would ignore the file. This library has also been updated to the
new validator location and the newest version of validator and the new
api for Valid.